changed argllint to support unsigned long int

+++ libflow.c
@@ -6,7 +6,7 @@
 /*   By: ocarlos- <ocarlos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/30 21:42:14 by ocarlos-          #+#    #+#             */
-/*   Updated: 2020/07/20 09:19:12 by ocarlos-         ###   ########.fr       */
+/*   Updated: 2020/07/20 20:47:57 by ocarlos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int		ft_vprintf(t_prtform *form, va_list lst, int count)
 	}
 	if (ft_testllint(form->typ))
 	{
-		argllint = va_arg(lst, unsigned int);
+		argllint = va_arg(lst, unsigned long int);
 		form->siz = ft_redllint(argllint, form);
 	}
 	if (form->typ == 's')
